job_hunted/requirements.txt
Django==2.0.6
django-cors-headers==2.2.0
djangorestframework==3.8.2
djangorestframework-jsonapi==2.0.0b2
inflection==0.3.1
pytz==2018.4
